refactor: recommend 패키지 구조 도메인 레이어로 변경
