+++ src/main/java/net/detalk/api/post/controller/ProductPostController.java
@@ -8,8 +8,8 @@ import lombok.RequiredArgsConstructor;
 import net.detalk.api.post.controller.request.UpdateProductPostRequest;
 import net.detalk.api.post.controller.response.CreateProductPostResponse;
 import net.detalk.api.post.controller.response.GetProductPostResponse;
-import net.detalk.api.controller.v1.request.CreateRecommendRequest;
-import net.detalk.api.service.RecommendService;
+import net.detalk.api.post.controller.request.CreateRecommendRequest;
+import net.detalk.api.post.service.RecommendService;
 import net.detalk.api.support.CursorPageData;
 import net.detalk.api.post.controller.request.CreateProductPostRequest;
 import net.detalk.api.post.service.ProductPostService;
