+++ src/test/java/net/detalk/api/service/ProductPostServiceTest.java
@@ -31,7 +31,7 @@ import net.detalk.api.post.domain.ProductPost;
 import net.detalk.api.post.domain.ProductPostSnapshot;
 import net.detalk.api.tag.domain.Tag;
 import net.detalk.api.domain.exception.DuplicateCreatePostException;
-import net.detalk.api.domain.exception.InvalidRecommendCountRequest;
+import net.detalk.api.post.domain.exception.InvalidRecommendCountRequest;
 import net.detalk.api.mock.FakeTimeHolder;
 import net.detalk.api.mock.FakeUUIDGenerator;
 import net.detalk.api.product.repository.ProductLinkRepository;
