+++ src/test/java/net/detalk/api/service/RecommendServiceTest.java
@@ -16,14 +16,15 @@ import java.time.Instant;
 import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Optional;
-import net.detalk.api.controller.v1.request.CreateRecommendRequest;
-import net.detalk.api.domain.Recommend;
-import net.detalk.api.domain.exception.DuplicateRecommendationException;
+import net.detalk.api.post.controller.request.CreateRecommendRequest;
+import net.detalk.api.post.domain.Recommend;
+import net.detalk.api.post.domain.exception.DuplicateRecommendationException;
 import net.detalk.api.post.domain.exception.ProductPostNotFoundException;
 import net.detalk.api.mock.FakeTimeHolder;
+import net.detalk.api.post.repository.RecommendProductRepository;
+import net.detalk.api.post.repository.RecommendRepository;
 import net.detalk.api.post.service.ProductPostService;
-import net.detalk.api.repository.RecommendProductRepository;
-import net.detalk.api.repository.RecommendRepository;
+import net.detalk.api.post.service.RecommendService;
 import net.detalk.api.support.TimeHolder;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
